Add a backgroundvideo Param, create video tag
If backgroundvideo is specified, the theme will load the video in the
background of the main portion of the page.

The video must be an mp4 file, and should be located in the same bundle
as the index page.

Couldn't quite get the height to render correctly, so a small amount of
fudging was required in the css